pyproject.toml
[project]
name = "repo-summary"
version = "0.1.0"
description = "Generate repository summaries from GitLab, GitHub, and other Git hosting services"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
]

[project.scripts]
repo-summary = "repo_summary.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []
